+++ src/service/middlewares/validation/index.ts
@@ -1,2 +1,2 @@
-export * from './validate-schema';
-export * from './validate-param-schema';
+export * from './validate-schema/validate-schema';
+export * from './validate-param-schema/validate-param-schema';
